feat: add stock API

+++ backend/src/modules/product/product.service.js
@@ -8,18 +8,18 @@ import { deleteImage, uploadImage } from '../../../services/storage.service.js';
 export const getAllProducts = async ({ categoryId, keyword }) => {
   const filter = { is_active: true };
   if (categoryId) {
-    filter.category_id = categoryId;
+    filter.category = categoryId;
   }
   if (keyword) {
     filter.name = { $regex: keyword, $options: 'i' };
   }
   return await Product.find(filter)
-    .populate('category_id', 'name')
+    .populate('category', 'name')
     .sort({ createdAt: -1 });
 };
 
 export const getProductById = async (id) => {
-  return await Product.findById(id).populate('category_id', 'name');
+  return await Product.findById(id).populate('category', 'name');
 };
 
 export const createProduct = async (payload, files = []) => {
@@ -32,7 +32,7 @@ export const createProduct = async (payload, files = []) => {
     });
   }
 
-  const categoryExists = await Category.exists({ _id: payload.category_id });
+  const categoryExists = await Category.exists({ _id: payload.category });
   if (!categoryExists) {
     throw new ErrorWithStatus({
       status: HTTP_STATUS.NOT_FOUND,
@@ -53,9 +53,7 @@ export const createProduct = async (payload, files = []) => {
       expiry_date: payload.expiry_date,
       weight: payload.weight,
       unit: payload.unit,
-      stock_quantity: payload.stock_quantity ?? 0,
-      is_active: payload.is_active ?? true,
-      category_id: payload.category_id,
+      category: payload.category,
       images: uploadedImages.map((key, index) => ({
         url: `${process.env.CLOUDFLARE_R2_PUBLIC_URL}/${key}`,
         publicId: key,
@@ -83,21 +81,41 @@ export const updateProduct = async (id, payload, files = []) => {
   if (files && files.length > 0) {
     const existingProduct = await Product.findById(id);
 
-    if (existingProduct.images?.length > 0) {
-      await Promise.all(
-        existingProduct.images.map((img) => deleteImage(img.publicId)),
-      );
+    if (!existingProduct) {
+      throw new ErrorWithStatus({
+        status: HTTP_STATUS.NOT_FOUND,
+        message: USERS_MESSAGES.PRODUCT_NOT_FOUND,
+      });
     }
 
     const uploadedImages = await Promise.all(
       files.map((file) => uploadImage(file, 'products')),
     );
 
+    if (existingProduct.images?.length > 0) {
+      await Promise.all(
+        existingProduct.images.map((img) => deleteImage(img.publicId)),
+      );
+    }
+
     payload.images = uploadedImages.map((key, index) => ({
       url: `${process.env.CLOUDFLARE_R2_PUBLIC_URL}/${key}`,
       publicId: key,
       isPrimary: index === 0,
     }));
+  }
+
+  if (payload.category) {
+    const categoryExists = await Category.exists({
+      _id: payload.category,
+    });
+
+    if (!categoryExists) {
+      throw new ErrorWithStatus({
+        status: HTTP_STATUS.NOT_FOUND,
+        message: 'Category not found',
+      });
+    }
   }
 
   const product = await Product.findByIdAndUpdate(id, payload, {
@@ -115,11 +133,7 @@ export const updateProduct = async (id, payload, files = []) => {
 };
 
 export const deleteProduct = async (id) => {
-  const product = await Product.findByIdAndUpdate(
-    id,
-    { is_active: false },
-    { new: true },
-  );
+  const product = await Product.findById(id);
 
   if (!product) {
     throw new ErrorWithStatus({
@@ -128,5 +142,14 @@ export const deleteProduct = async (id) => {
     });
   }
 
+  if (product.stock_quantity > 0) {
+    throw new ErrorWithStatus({
+      status: HTTP_STATUS.BAD_REQUEST,
+      message: 'Cannot delete product with remaining stock',
+    });
+  }
+
+  product.is_active = false;
+  await product.save();
   return product;
 };
